[CHANGE] rename manager
